Initial codebase includes using POST to upload and GET to download and delete!
Begin rewrite of A-Share,

To upload,
POST with form-data for the file to upload

To download,
GET /files/<file>

To delete,
GET /delete/<token>

--
Files are received in `/uploads/temp` and then moved to `/uploads` with a generated discriminator.

The files in `/uploads` are given a token for deletion in `/registry` to ensure recipients may not destroy content that is intended to be seen more than once.